+++ packages/vendure-plugin-accept-blue/src/service/accept-blue-service.ts
@@ -251,7 +251,7 @@ export class AcceptBlueService implements OnApplicationBootstrap {
       transactionId: chargeTransactionId
         ? String(chargeTransactionId)
         : undefined,
-      metadata: chargeTransaction,
+      metadata: chargeTransaction || {},
     };
   }